Split docker package into several files
